regression-tests/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=-
#
appium-python-client==3.2.1
    # via robotframework-appiumlibrary
attrs==25.3.0
    # via
    #   jsonschema
    #   outcome
    #   referencing
    #   trio
beautifulsoup4==4.13.4
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   requests
    #   selenium
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-click
    #   robotframework-browser
    #   robotframework-tidy
    #   typer
colorama==0.4.6
    # via robotframework-tidy
decorator==5.2.1
    # via robotframework-appiumlibrary
docutils==0.21.2
    # via robotframework-appiumlibrary
grpcio==1.71.0
    # via
    #   grpcio-tools
    #   robotframework-browser
grpcio-tools==1.71.0
    # via robotframework-browser
h11==0.16.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   trio
isort==6.0.1
    # via -r requirements.in
jinja2==3.1.6
    # via
    #   robotframework-robocop
    #   robotframework-tidy
jsonpath-ng==1.7.0
    # via robotframework-jsonlibrary
jsonschema==4.23.0
    # via robotframework-jsonlibrary
jsonschema-specifications==2024.10.1
    # via jsonschema
kitchen==1.2.6
    # via robotframework-appiumlibrary
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
natsort==8.4.0
    # via
    #   robotframework-pabot
    #   seedir
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
overrides==7.7.0
    # via robotframework-browser
pathspec==0.12.1
    # via
    #   robotframework-robocop
    #   robotframework-tidy
platformdirs==4.3.7
    # via robotframework-robocop
ply==3.11
    # via jsonpath-ng
protobuf==5.29.3
    # via
    #   -r requirements.in
    #   grpcio-tools
    #   robotframework-browser
pygments==2.19.1
    # via rich
pysocks==1.7.1
    # via urllib3
pytz==2025.1
    # via robotframework-robocop
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via robotframework-requests
rich==13.9.4
    # via
    #   rich-click
    #   typer
rich-click==1.8.5
    # via robotframework-tidy
robotframework==7.2.2
    # via
    #   -r requirements.in
    #   robotframework-appiumlibrary
    #   robotframework-assertion-engine
    #   robotframework-browser
    #   robotframework-jsonlibrary
    #   robotframework-pabot
    #   robotframework-requests
    #   robotframework-robocop
    #   robotframework-stacktrace
    #   robotframework-tidy
robotframework-appiumlibrary==2.1.0
    # via -r requirements.in
robotframework-assertion-engine==3.0.3
    # via
    #   -r requirements.in
    #   robotframework-browser
robotframework-browser==19.5.1
    # via -r requirements.in
robotframework-jsonlibrary==0.5
    # via -r requirements.in
robotframework-pabot==4.2.0
    # via -r requirements.in
robotframework-pythonlibcore==4.4.1
    # via
    #   -r requirements.in
    #   robotframework-assertion-engine
    #   robotframework-browser
robotframework-requests==0.9.7
    # via -r requirements.in
robotframework-robocop==6.1.1
    # via -r requirements.in
robotframework-stacktrace==0.4.1
    # via robotframework-pabot
robotframework-tidy==4.16.0
    # via -r requirements.in
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
ruff==0.11.9
    # via -r requirements.in
seedir==0.5.0
    # via robotframework-browser
selenium==4.16.0
    # via
    #   appium-python-client
    #   robotframework-appiumlibrary
shellingham==1.5.4
    # via typer
six==1.17.0
    # via robotframework-appiumlibrary
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.6
    # via beautifulsoup4
tomli==2.2.1
    # via robotframework-tidy
tomli-w==1.2.0
    # via robotframework-robocop
trio==0.29.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typer==0.15.3
    # via robotframework-robocop
typing-extensions==4.12.2
    # via
    #   beautifulsoup4
    #   referencing
    #   rich-click
    #   typer
urllib3[socks]==2.3.0
    # via
    #   requests
    #   selenium
wrapt==1.17.2
    # via robotframework-browser
wsproto==1.2.0
    # via trio-websocket

# The following packages are considered to be unsafe in a requirements file:
# setuptools
